Bool Coercion and String Split SMTs
[Card](https://rentpath.atlassian.net/browse/SRV-1179)

This adds two new SMTs (Single Message Transforms) that are effectively the inversions of the original two.

`BoolCoercionTransform` attempts to interpret a non-boolean source value as a boolean, e.g.:
- String values "true", "yes", and "1" (case-insensitive) are interpreted as logical `true`
- String values "false", "no", and "0" (case-insensitive) are interpreted as logical `false`
- Integer value 0 is interpreted as logical `false`
- Any other integer value is interpreted as logical `true`
- Boolean values are unchanged
Additionally, this transform has an optional flag to interpret a null source value as logical `false`.

`StringSplitTransform` attempts to interpret a delimited-string source value as a list of strings, with options to interpret null and empty values appropriately (see config class `StringSplitTransformConfig` for details).