Add union resolver SMT
[Card](https://rentpath.atlassian.net/browse/SRV-5038)

The Avro converter that handles translating an Avro schema to a connect schema is not intelligent enough to appropriately handle either-or unions, such as when a given field may be represented as multiple numeric types. This allows the specification of a transform that will select one of the specified numeric types in order of priority as configured.